Modified classes to match the spec
