+++ AdminSystem/StaffConfirmDelete.aspx
@@ -1,34 +1,18 @@
-<%@ Page Language="C#" AutoEventWireup="true" CodeFile="StaffConfirmDelete.aspx.cs" Inherits="StaffConfirmDelete" %>
-
-<!-- 
-This ASPX was generated by a purpose built internal Monday 9AM Ensemble development utility.
-The HTML content was generated by bootstrap studio.
-
-It looks dreadful in the Visual Studio preview, but fine when hosted in a browser.
-The bootstrap front end seems perfectly compatable.
--->
-
+﻿<%@ Page Language="C#" AutoEventWireup="true" CodeFile="StaffConfirmDelete.aspx.cs" Inherits="StaffConfirmDelete" %>
 
 <!DOCTYPE html>
-<html style="height: 100%;width: 100%;">
 
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, shrink-to-fit=no">
-    <title>LS Goods - Dylan</title>
-    <link rel="stylesheet" href="assets/bootstrap/css/bootstrap.min.css">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/3.5.2/animate.min.css">
-    <link rel="stylesheet" href="assets/css/styles.css">
+<html xmlns="http://www.w3.org/1999/xhtml">
+<head runat="server">
+    <title></title>
 </head>
-
-<body style="height: 100%;width: 100%;background: var(--dark);">
-    <div class="container d-xl-flex justify-content-xl-center align-items-xl-center rubberBand animated" style="width: 100%;height: 100%;background: var(--cyan);">
-        <form method="post" runat="server"><asp:Label ID="lblLogin" runat="server" Text="Login"></asp:Label><asp:TextBox runat="server" ID="txtUserID">Enter your ID</asp:TextBox><asp:TextBox runat="server" ID="txtPasswrod">Enter your password</asp:TextBox>
-            <asp:CheckBox ID="chkRememberMe" runat="server" Text="Remember my ID" OnCheckedChanged="chkRememberMe_CheckedChanged"/><asp:Button ID="btnEnter" runat="server" Text="Enter"  OnClick="btnEnter_Click"></asp:Button>
-        </form>
-    </div>
-    <script src="assets/js/jquery.min.js"></script>
-    <script src="assets/bootstrap/js/bootstrap.min.js"></script>
+<body>
+    <form id="form1" runat="server">
+        <asp:Label ID="lblDeleteRecord" runat="server" Text="Are you sure you want to delete this record?"></asp:Label>
+        <div>
+        </div>
+        <asp:Button ID="btnYes" runat="server" OnClick="btnYes_Click" Text="Yes" />
+        <asp:Button ID="btnNo" runat="server" OnClick="btnNo_Click" Text="No" />
+    </form>
 </body>
-
 </html>
